Massive changes: new interpolation methods and refactoring (#5)
1. Removed the `Polynomial` class.
2. Added `alfi::poly::val` functions for polynomial evaluation.
3. Introduced `alfi::poly::lagrange` for Lagrange polynomial interpolation.
4. Introduced `alfi::poly::newton` for Newton polynomial interpolation.
5. Implemented `alfi::misc::barycentric` for barycentric interpolation
   (adapted from https://people.maths.ox.ac.uk/trefethen/barycentric.pdf).

6. Added basic tests for Lagrange, Newton, and barycentric interpolation methods.
7. Created an example application in `examples` directory
   to demonstrate and explore the implemented interpolation methods.

8. Added `config.h` for configuring default number and container types.
9. Enabled container type configuration for all functions.

10. Added optimization flags for applications in the `examples` directory.

+++ examples/dist_QCustomPlot/dist_QCustomPlot.cpp
@@ -200,7 +200,7 @@ private:
 
 		double current_layer = -1;
 
-		for (int i = 0; i < distributions.size(); ++i) {
+		for (size_t i = 0; i < distributions.size(); ++i) {
 			const auto& [name, x_data] = distributions[i];
 			QVector<double> x, y;
 			for (const double& value : x_data) {
